requirements.txt
cdsapi
xarray
numpy
pandas
matplotlib
sentinelsat
rasterio
geopandas
requests
h5netcdf
netCDF4
folium
geopandas
shapely
python-dotenv
